fix(network-automation): let the auto-import summary test run in CI
Not part of the Security Events work — master's App Test job has been red
since #3436 landed, and this branch inherited that failure when it merged
master in, so the fix rides along here to unblock it.

App/Tests/Dashboard/RunAutoImportRuleModal.test.ts imported
describeAutoImportRun from RunAutoImportRuleModal.tsx. The function is
pure string formatting, but living in a .tsx put react in its import
graph — and react is a Dashboard dependency that App's own install never
provides, so in CI the suite could not resolve it and failed to run at
all. It passes locally only where a Dashboard node_modules happens to
exist. This is the same constraint that makes the other tests in
App/Tests/Dashboard read .tsx sources as text rather than import them.

Moved the helper and its two formatters into a react-free sibling,
AutoImportRunSummary.ts; the modal imports it, the test imports it
directly, and the test's Common/UI/Config mock — which existed only to
survive the modal's browser import graph — is gone with it.

Verified by moving the local Dashboard node_modules aside to reproduce
CI's environment: the suite passes, and the full App suite reports 11137
tests passing with no test failures.
